fix build and ci dependencies, compile error and warning on ubuntu 18

+++ src/qtgui/plotter.cpp
@@ -1695,7 +1695,7 @@ void CPlotter::moveToCenterFreq()
 /** Center FFT plot around the demodulator frequency. */
 void CPlotter::moveToDemodFreq(size_t idx)
 {
-    if (idx >= m_demod.size()) {
+    if ((int)idx >= m_demod.size()) {
         qInfo() << "CPlotter::moveToDemodFreq invalid idx" << idx;
     }
